refactor: diplomacy wave 5 slice E test support densify (#4130)
Densify evidence_rules_scenarios and CTA fixtures below the 400-line
support cap, ratchet diplomacy test-support LOC to 2324, and split four
near-cap diplomacy test suites into focused runners.
