+++ SPEC/program/repo-lint.md
@@ -204,11 +204,11 @@
 | `tool/check_diplomacy_test_duplicate_descriptions.dart` | Fail when the same single-line `test('...')` / `testWidgets('...')` description appears in more than one file under `packages/colonizethis_diplomacy/test/**` (Refs #3825); rule `repo.diplomacy_test_duplicate_descriptions` |
 | `tool/check_diplomacy_test_inline_game_ctor.dart` | Forbid inline `Game(` constructor calls under **all** `packages/colonizethis_diplomacy/test/**` Dart files (exceptions-only allowlist; empty after wave-3 fixture migration). Suites MUST build games through shared helpers in `package:colonizethis_diplomacy_test_support` (Refs #3825, #3837, #4028); rule `repo.diplomacy_test_inline_game_ctor` |
 | `tool/check_diplomacy_relation_upsert_commit.dart` | Forbid raw `diplomacyRelations: relationsIndex.toList()` under `packages/colonizethis_diplomacy/lib/**` outside `diplomacy_relation_upsert.dart`; commits MUST use `committedRelations` / `withCommittedRelations` (Refs #4028); rule `repo.diplomacy_relation_upsert_commit` |
-| `tool/check_diplomacy_test_support_loc.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy_test_support/lib/**` physical Dart LOC (≤2488 after wave-4 densify; ≤2500 band met); rule `repo.diplomacy_test_support_loc` (Refs #4028, #4037) |
+| `tool/check_diplomacy_test_support_loc.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy_test_support/lib/**` physical Dart LOC (≤2324 after wave-5 densify; was ≤2488 after wave-4); rule `repo.diplomacy_test_support_loc` (Refs #4028, #4037, #4130) |
 | `tool/check_diplomacy_test_no_src_imports.dart` | Forbid `package:colonizethis_diplomacy/src/` imports under `packages/colonizethis_diplomacy/test/**` except the allowlisted `diplomacy_phase_types_split_test.dart` deep-import gate; tests MUST use the public barrel or `colonizethis_diplomacy_test_support` (Refs #3837); rule `repo.diplomacy_test_no_src_imports` |
 | `tool/check_diplomacy_test_file_size.dart` | Fail when any `.dart` file under `packages/colonizethis_diplomacy/test/**` exceeds 400 physical lines (aligns with economy test file cap); rule `repo.diplomacy_test_file_size` |
 | `tool/check_diplomacy_lib_file_size.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy/lib/**` physical lines per file (wave-5 ≤400; shrink-only grandfather allowlist for two resolver lib hotspots during transition; generated suffixes excluded); rule `repo.colonizethis_diplomacy_lib_file_size` (Refs #4130) |
-| `tool/check_diplomacy_test_support_file_size.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy_test_support/lib/**` physical lines per file (wave-5 ≤400; shrink-only grandfather allowlist for `evidence_rules_scenarios.dart` during densify; generated suffixes excluded); rule `repo.colonizethis_diplomacy_test_support_file_size` (Refs #4130) |
+| `tool/check_diplomacy_test_support_file_size.dart` | Ratchet ceiling for `packages/colonizethis_diplomacy_test_support/lib/**` physical lines per file (wave-5 ≤400; shrink-only grandfather allowlist empty after slice E densify; generated suffixes excluded); rule `repo.colonizethis_diplomacy_test_support_file_size` (Refs #4130) |
 | `tool/check_diplomacy_test_support_package_only.dart` | Forbid package-local fixture modules under `packages/colonizethis_diplomacy/test/support/**` and new `*_test_support.dart` shims under `packages/colonizethis_diplomacy/test/**`; shared fixtures MUST live in `colonizethis_diplomacy_test_support` (Refs #3837); rule `repo.diplomacy_test_support_package_only` |
 | `tool/check_ai_dedup_weight_scale_clamp.dart` | Forbid the inline weight-scale clamp idiom (`<= 0.0) { return 0; }` guard + `final clamped = X > 1.0 ? 1.0 : X;` + `return (Y * clamped).round();`) under `packages/colonizethis_ai/lib/**` outside the canonical helper file `src/planning/planning_helpers.dart`; filter/scoring code MUST call the shared `scaleWeightedBonus(weight, baseConstant)` helper (Refs #3278); rule `repo.ai_dedup_weight_scale_clamp` |
 | `tool/check_ai_dedup_colonial_pressure_scale.dart` | Forbid inline colonial-pressure null-fallback ternaries (`… ? 1.0 : 0.0`) under `packages/colonizethis_ai/lib/**` outside `src/planning/planning_helpers.dart`; planner code MUST call the shared `colonialPressureScaleFromWeight` helper (Refs #3822); rule `repo.ai_dedup_colonial_pressure_scale` |
@@ -383,7 +383,7 @@ Large `app/test/**` widget/screen suites that are split into numbered fragments 
 
 **Diplomacy lib file size (Refs #4130):** `repo.colonizethis_diplomacy_lib_file_size` fails when any non-generated Dart file under `packages/colonizethis_diplomacy/lib/**` exceeds **400** physical lines outside the shrink-only grandfather allowlist (two resolver lib hotspots — `faction_absorption_engine.dart` and `war_resolver.dart` — during wave-5 transition slices; dossier dialogue/evidence and overture/CTA resolver modules split in slices B and D).
 
-**Diplomacy test support file size (Refs #4130):** `repo.colonizethis_diplomacy_test_support_file_size` fails when any non-generated Dart file under `packages/colonizethis_diplomacy_test_support/lib/**` exceeds **400** physical lines outside the shrink-only grandfather allowlist (`evidence_rules_scenarios.dart` during densify slices).
+**Diplomacy test support file size (Refs #4130):** `repo.colonizethis_diplomacy_test_support_file_size` fails when any non-generated Dart file under `packages/colonizethis_diplomacy_test_support/lib/**` exceeds **400** physical lines outside the shrink-only grandfather allowlist (empty after wave-5 slice E densify).
 
 **Turn test support LOC (Refs #4039, #4113):** `repo.turn_test_support_loc` fails when physical LOC of all Dart under `packages/colonizethis_turn/test/support/` exceeds the documented post-densify ceiling (**5800**). Ratchet downward as densify slices shrink the tree.
